Added sending list of metrics in pieces.
If the list of metrics is too large (more than 250 items) it will be
divided into pieces each of which will be sent separately. The idea and
the specific number of metrics taken from the original implementation of
zabbix_sender.

The number of metrics on one piece can be optionally specified when
creating the object ZabbixSender.