fix ipv4/ipv6 fallback
The current fallback to ipv6 won't work in most
cases (successful socket.socket(socket.AF_INET)
doesn't means ipv4 connectivity on the remote host).

What we are trying to do, is to connect to the right IP
(v4/v6) depending on DNS records (A / AAAA) and system
preference and it's already done system-wide via
getaddrinfo.